+++ platforms/platform-facebookmessenger/src/FacebookMessengerUser.ts
@@ -1,12 +1,12 @@
 import { JovoUser } from '@jovotech/framework';
 import { FacebookMessengerResponse } from '@jovotech/output-facebookmessenger';
 import { FacebookMessengerRequest } from './FacebookMessengerRequest';
-import { MessengerBot } from './MessengerBot';
+import { FacebookMessenger } from './FacebookMessenger';
 
 export class FacebookMessengerUser extends JovoUser<
   FacebookMessengerRequest,
   FacebookMessengerResponse,
-  MessengerBot
+  FacebookMessenger
 > {
   get id(): string {
     return this.jovo.$request.messaging?.[0]?.sender?.id || 'FacebookMessengerUser';
